Add files via upload
Unsure what to put here - For clarity sake, this is nothing more than toying around to see what happens in tkinter. Nothing about this is suggested. It's the result of toying around for 2 hours in a language I don't use, and with a module I'm unfamiliar with. Be advised. Do not use as template.